fix(*): moved utils to other pkg
