+++ pkg/util/net/ips.go
@@ -32,3 +32,16 @@ func ToV6(ip string) string {
 	}
 	return ip
 }
+
+func IsAddressIPv6(address string) bool {
+	if address == "" {
+		return false
+	}
+
+	ip := net.ParseIP(address)
+	if ip == nil {
+		return false
+	}
+
+	return ip.To4() == nil
+}
