+++ pkg/util/net/ips_test.go
@@ -16,3 +16,12 @@ var _ = DescribeTable("ToV6",
 	Entry("v4 adds prefix", "240.0.0.0", "::ffff:f000:0"),
 	Entry("v4 adds prefix", "240.0.255.0", "::ffff:f000:ff00"),
 )
+
+var _ = DescribeTable("IsIPv6",
+	func(given string, expected bool) {
+		Expect(net.IsAddressIPv6(given)).To(Equal(expected))
+	},
+	Entry("127.0.0.1 should not be IPv6 ", "127.0.0.1", false),
+	Entry("should be IPv6", "2001:0db8:0000:0000:0000:ff00:0042:8329", true),
+	Entry("::6", "::6", true),
+)
