--- pkg/xds/generator/utils.go
@@ -1,16 +1,0 @@
-package generator
-
-import "net"
-
-func IsAddressIPv6(address string) bool {
-	if address == "" {
-		return false
-	}
-
-	ip := net.ParseIP(address)
-	if ip == nil {
-		return false
-	}
-
-	return ip.To4() == nil
-}
